+++ app/components/MobileMenu.tsx
@@ -2,16 +2,14 @@
 
 import dynamic from 'next/dynamic'
 import Link from 'next/link'
-import { useEffect, useState } from 'react'
+import { useState } from 'react'
 import { NAV_ROUTES } from '@/utils/constants'
 import EyeJayEsLogo from '@/lib/EyeJayEsLogo'
-import { throttle } from '@/utils/helpers'
 
 const DarkModeSwitch = dynamic(() => import('@/lib/DarkModeSwitch'), { ssr: false })
 
 export default function MobileMenu() {
 	const [isOpen, setIsOpen] = useState<boolean>(false)
-	const [isTransparent, setIsTransparent] = useState<boolean>(true)
 
 	const handleHamburgerClick = () => {
 		setIsOpen((prevIsOpen) => {
@@ -27,25 +25,9 @@ export default function MobileMenu() {
 		}, 200)
 	}
 
-	useEffect(() => {
-		const handleScroll = throttle(() => {
-			setIsTransparent(window.scrollY < 72)
-		}, 150)
-
-		window.addEventListener('scroll', handleScroll)
-
-		return () => {
-			window.removeEventListener('scroll', handleScroll)
-		}
-	}, [])
-
 	return (
-		<div
-			className={`md:hidden transition-colors fixed top-0 left-0 z-50 w-full h-[72px] ${
-				isTransparent ? 'bg-transparent' : 'shadow-md bg-white dark:bg-grey-blue'
-			}`}
-		>
-			<div className={`fixed inset-0 h-full ${isOpen ? '' : ' pointer-events-none'}`}>
+		<div className="md:hidden h-full">
+			<div className={`fixed inset-0 w-full h-lvh ${isOpen ? '' : ' pointer-events-none'}`}>
 				<div
 					className={`transition-all absolute inset-0 flex flex-col justify-end pb-6 pl-6 bg-teal-100 dark:bg-grey-blue ${
 						isOpen ? 'opacity-1' : 'opacity-0'
@@ -66,7 +48,7 @@ export default function MobileMenu() {
 					<DarkModeSwitch />
 				</div>
 				<Link
-					className="transition-colors absolute top-[22px] left-4 flex items-center gap-3 text-lg font-bold text-indigo dark:text-white"
+					className="transition-colors absolute top-[18px] left-4 flex items-center gap-3 text-lg font-bold text-indigo dark:text-white"
 					href="/"
 				>
 					<EyeJayEsLogo width={80} />
@@ -76,7 +58,7 @@ export default function MobileMenu() {
 			<button
 				aria-label="Mobile navigation"
 				type="button"
-				className="transition-colors fixed top-4 right-4 flex justify-center items-center gap-2 h-10 w-10 rounded-lg dark:focus:outline-none dark:focus:ring-1 dark:focus:ring-gray-600 text-indigo dark:text-white"
+				className="transition-colors fixed top-3 right-4 flex justify-center items-center gap-2 h-10 w-10 rounded-lg dark:focus:outline-none dark:focus:ring-1 dark:focus:ring-gray-600 text-indigo dark:text-white"
 				onClick={handleHamburgerClick}
 			>
 				<svg
